Use simple list instead of set for the feeds/entries lists.
I've used set() so far to guarantee uniqueness of the elements, but it
causes a test-fail in modeltest sometimes as the order of a set is not
guaranteed.

As I don't yet care too much about uniqueness of the elements lets resort
to a simple list for now. I'm not even sure anymore wether its really
necessary to have the actual lists unique or wether the uniqueness-check
should be done elsewhere.