Support configuration of the systray color and font for the unread text
With this change a user can configure the font and font-color to be used
to draw the number of unread articles in the systray icon. That makes it
easy to adjust the texts look to the users liking and to possible
readability problems when contrast is too low or the font is otherwise
hard to read.

fixes #12